requirements.txt
annotated-types==0.7.0
anthropic==0.44.0
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.43
langchain-deepseek==0.1.2
langchain-openai==0.3.8
langgraph==0.2.66
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.0
loguru
msgpack
newsapi-python
openai
orjson
pydantic=
pydantic_core
python-dotenv
PyYAML
regex
requests
requests-toolbelt
six
sniffio
tavily-python
tenacity
tiktoken
tqdm
urllib3
zstandard
